Se organiza el CaptureColor.py para cada juego. se eliminan archios inecesarios. Se calibran colores
