+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/Python-Games-master.iml" filepath="$PROJECT_DIR$/.idea/Python-Games-master.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/ProyectoJuegoPDI.iml" filepath="$PROJECT_DIR$/.idea/ProyectoJuegoPDI.iml" />
     </modules>
   </component>
 </project>
