--- JuegoTetrisPDI/.idea/.gitignore
@@ -1,8 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-# Datasource local storage ignored files
-/../../../../../../../../../../../../:\Users\Luis Felipe Cadavid\Documents\ACADEMIA\UNIVERSIDAD DE ANTIOQUIA\SEMESTRES\2020-1\PROC. DIGITAL DE IMÁGENES\PROYECTO JUEGO\Python-Games-master\.idea/dataSources/
-/dataSources.local.xml
-# Editor-based HTTP Client requests
-/httpRequests/
